Revert "feat: add testing"
