+++ src/app/app.component.html
@@ -1,3 +1,3 @@
-<section class="app">
-    <router-outlet></router-outlet>
-</section>
+<h1>Hello Baton Candidate!</h1>
+
+<router-outlet></router-outlet>
