--- src/app/app.component.scss
@@ -1,3 +1,0 @@
-.app {
-    background-color: #f0f0f0;
-}
