--- src/app/shared/components/todo/todo.component.scss
@@ -1,23 +1,0 @@
-.todo {
-    background-color: #fff;
-    box-shadow: 0 1px 5px -2px grey;
-    margin: 1vh 0;
-    padding: 2vh 3vh;
-    font-size: 2vh;
-    border-radius: 2px;
-
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-
-    &__btn {
-        padding: 0.5vh 2vh;
-        font-size: 15px;
-        font-weight: 600;
-        background-color: #0078e7;
-        color: #fff;
-        box-shadow: none;
-        border: none;
-        border-radius: 4px;
-    }
-}
