--- src/app/todo-list/done/done.component.html
@@ -1,9 +1,0 @@
-<section class="todo">
-    <h1 class="todo--title">Todo</h1>
-    <button (click)="navigateBack()">&#8592; Back</button>
-    <ul class="todo__list">
-        <li class="todo__list--item" *ngFor="let todo of todos">
-            <app-todo [todo]="todo"></app-todo>
-        </li>
-    </ul>
-</section>
