--- src/app/todo-list/done/done.component.scss
@@ -1,18 +1,0 @@
-.todo {
-    width: 100vh;
-    margin: 0 auto;
-    padding: 5vh 0;
-    height: 100%;
-
-    &--title {
-        font-size: xx-large;
-        text-align: center;
-        margin-top: 0;
-    }
-
-    &__list {
-        list-style: none;
-        margin: 0;
-        padding: 0;
-    }
-}
