--- src/app/todo-list/todo-list.component.html
@@ -1,1 +1,0 @@
-<router-outlet></router-outlet>
