--- src/app/todo-list/undone/undone.component.html
@@ -1,9 +1,0 @@
-<section class="todo">
-    <h1 class="todo--title">Todo</h1>
-    <button (click)="navigateDone()">View Done Todos &#8594;</button>
-    <ul class="todo__list">
-        <li class="todo__list--item" *ngFor="let todo of todos">
-            <app-todo [todo]="todo" (done)="onDone(todo)"></app-todo>
-        </li>
-    </ul>
-</section>
