+++ src/environments/environment.prod.ts
@@ -1,4 +1,3 @@
 export const environment = {
-  production: true,
-  API_URL: 'https://jsonplaceholder.typicode.com'
+  production: true
 };
