+++ src/environments/environment.ts
@@ -3,8 +3,7 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  API_URL: 'https://jsonplaceholder.typicode.com'
+  production: false
 };
 
 /*
